Moving Cacher back to nodes.util.
