+++ src/main/scala/workflow/ExtractSaveablePrefixes.scala
@@ -1,5 +1,7 @@
 package workflow
 
+import nodes.util.Cacher
+
 /**
  * Extract the prefixes of all Nodes whose state we want to save for reuse by other Pipeline apply and fit calls.
  * This is all nodes that either have a Cacher or an EstimatorOperator as the internal operator.
